Error de editar hoteles arreglado

+++ src/app/components/hoteles/hoteles.component.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, inject, Input } from '@angular/core';
 import { Hoteles } from '../../model/hoteles';
 import { HotelesService } from '../../services/hoteles.service';
+import { EmpleadosService } from '../../services/empleados.service';
 import { RouterLink } from '@angular/router';
 
 @Component({
@@ -15,6 +16,7 @@ export class HotelesComponent {
   @Input() hotelesList: Hoteles[] = [];
 
   hotelesService: HotelesService = inject(HotelesService);
+  empleadoService: EmpleadosService = inject(EmpleadosService);
 
   constructor() {
 
@@ -29,15 +31,25 @@ export class HotelesComponent {
   }
 
   deleteHotel(id: string) {
-    this.hotelesService.deleteHotel(id).subscribe(
-      
-      (response) => {
-        console.log('Hotel eliminado correctamente:', response);
-        this.hotelesList = this.hotelesList.filter((hotel) => hotel.id !== id);
+
+    this.empleadoService.getEmpleadoByHotelId(id).subscribe(
+      (empleados) => {
+        if (empleados.length > 0) {
+          alert('No se puede eliminar el hotel porque tiene empleados asociados');
+        } else {
+          this.hotelesService.deleteHotel(id).subscribe(
+            (response) => {
+              console.log('Hotel eliminado correctamente:', response);
+              this.hotelesList = this.hotelesList.filter((hotel) => hotel.id !== id);
+            },
+            (error) => {
+              console.error('Error al eliminar el hotel', error);
+            }
+          );
+        }
       },
       (error) => {
-        console.error('Error al eliminar el hotel', error);
-      }
-    );
+        console.error('Error al verificar empleados asociados', error);
+      });
   }
 }
